Add the Non-Homework Screen (#16)
* The Non HW screen

The file for creating the nonhomework screen. It allows you to add extracurriculars to the calendar.

* Changed the formating

* Update add_non_hw_screen.tsx

* Add Non-Homework screen(extracurriculars)

* Updated handleConfirm

Gave the date parameter in handleConfirm a Date type.

* Update handleConfirm

Changed the handleConfirm parameter to "date: Date" to make it have type Date

* Updated const demoValueControlledTextArea

Updated the const demoValueControlledTextArea to give parameter e "NativeSyntheticEvent<TextInputChangeEventData>". Also, updated the imports and removed any not being used and added NativeSyntheticEvent, and TextInputChangeEventData.