Modify Non_HW screen (#26)
* The Non HW screen

The file for creating the nonhomework screen. It allows you to add extracurriculars to the calendar.

* Changed the formating

* Update add_non_hw_screen.tsx

* Add Non-Homework screen(extracurriculars)

* Updated handleConfirm

Gave the date parameter in handleConfirm a Date type.

* Update handleConfirm

Changed the handleConfirm parameter to "date: Date" to make it have type Date

* Updated const demoValueControlledTextArea

Updated the const demoValueControlledTextArea to give parameter e "NativeSyntheticEvent<TextInputChangeEventData>". Also, updated the imports and removed any not being used and added NativeSyntheticEvent, and TextInputChangeEventData.

* Date and Name edit

Changed the name of functions. Also, the date picker is replaced with DateTime to add the date and time at once.

* Name Change

Edited the names for the extracurricular screen

* Changed to Prettier format

* Change to account for eslint

* Update main_stack.tsx